addeed error_msg module, added kanji functionality
